refactor(subnet_service): split monolithic lib.rs into focused modules
- Extract core types (SubnetId, SubnetEvent, etc.) into subnet.rs
- Extract SubnetService struct and lifecycle into service.rs
- Extract scoring helpers into scoring.rs
- Simplify run() with separate loops for clarity
- Remove unused test_tracker() function
- lib.rs now contains only re-exports
